# packages/schemas/recipe/openapi.yaml
...
openapi: 3.0.3
info:
  title: Commise Recipe API
  version: 1.0.0
  description: |-
    The recipe service owns this contract. It is DERIVED from the service’s authored zod
    (`src/**/*.schema.ts` plus the `@kitchensink/recipe-core` schemas those compose), never
    hand-written, and it is not a code-generation input.

    Every path is also served under a deprecated bare `/v1/...` alias for consumers configured
    outside this repo and for already-shipped mobile binaries (ADR-0011). New integrators must use
    the `/api/v1/...` spelling published here.

    Coverage is partial and the gaps are enumerated by the generator on every run: an operation with
    no response schema is one whose body has no authored zod yet, not one whose body is empty.
servers:
  - url: https://recipe.commise.app
    description: Production
  - url: https://recipe.sandbox.commise.app
    description: Sandbox
security:
  - clerkSessionToken: []
paths:
...
  /api/v1/account/export:
    get:
      operationId: exportAccount
      summary: Export everything the recipe service holds for the caller.
      security:
        - clerkSessionToken: []
      responses:
        "200":
          description: The caller’s full export.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountExport"
        "400":
          description: "The request body, query or path was rejected. `code` says by what: `VALIDATION_FAILED` from the boundary
            parser (with `details.fields` naming each offending field, including a rejected unknown key), or a domain
            code such as `INVALID_VISIBILITY` / `UNKNOWN_INGREDIENT`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No, invalid, expired, or wrong-authorized-party bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: "The caller has an in-flight account erasure, so writes are locked (HAZ-052). `code: ERASURE_IN_PROGRESS`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: "Per-user rate limit exceeded (`code: TOO_MANY_REQUESTS`). ⚠️ This body used to be a bare JSON STRING — the
            throttler exception passed through the filter unchanged — and is now the envelope."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          headers:
            Retry-After:
              description: Seconds until the limit resets.
              schema:
                type: integer
                minimum: 0
                exclusiveMinimum: true
        "500":
          description: Unexpected server error. The body never carries internals.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      description: "GDPR Art. 15 (access) + Art. 20 (portability). The read-only INVERSE of erasure: it returns the same
        owner-scoped roots erasure removes or keeps. Scoped ENTIRELY by the verified token — there is no body or query
        parameter that can point it at another account. Two shape decisions differ from the API’s read contracts on
        purpose: an absent value is an explicit `null` rather than an omitted key, and a `numeric` column
        (`averageRating`) stays a STRING, because a portability document reports what is stored. Photo and version
        entries carry METADATA and resolved CDN URLs, never bytes. Rate-limited with the tightest (export) cap."
  /api/v1/account/erasure:
    post:
      operationId: requestAccountErasure
      summary: Request irreversible erasure of the caller’s account data.
      security:
        - clerkSessionToken: []
      responses:
        "202":
          description: The queued or already-running erasure job.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErasureRequestAcceptedResponse"
        "400":
          description: "The request body, query or path was rejected. `code` says by what: `VALIDATION_FAILED` from the boundary
            parser (with `details.fields` naming each offending field, including a rejected unknown key), or a domain
            code such as `INVALID_VISIBILITY` / `UNKNOWN_INGREDIENT`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No, invalid, expired, or wrong-authorized-party bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "410":
          description: The account was already erased.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: "The caller has an in-flight account erasure, so writes are locked (HAZ-052). `code: ERASURE_IN_PROGRESS`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: "Per-user rate limit exceeded (`code: TOO_MANY_REQUESTS`). ⚠️ This body used to be a bare JSON STRING — the
            throttler exception passed through the filter unchanged — and is now the envelope."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          headers:
            Retry-After:
              description: Seconds until the limit resets.
              schema:
                type: integer
                minimum: 0
                exclusiveMinimum: true
        "500":
          description: Unexpected server error. The body never carries internals.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      description: "IRREVERSIBLE. `confirmationPhrase` is the intent gate and is REQUIRED; its VALUE is checked by the
        service, not by request validation, so the accepted literal is deliberately NOT published as an enum here — a
        `400` reports a mismatch without echoing what was expected. There is no `ownerId` field: the owner comes from
        the verified token, and the body is strict, so a smuggled one is rejected with a 400 rather than stripped —
        which matters here because the operation is irreversible, and it applies equally to a misspelled
        `publishRecipeIds`, whose silent loss would have removed recipes the caller elected to donate. Idempotent per
        C-007 — an in-flight job is RETURNED rather than a second one started, and there is no `409`, because a
        duplicate erasure request is the same request again. `publishRecipeIds` is the optional per-recipe DONATE
        election (CR-002 / U3b): a listed recipe is published and KEPT under a pseudonymous handle the erased owner no
        longer controls, instead of being removed; absent or empty means donate nothing."
      requestBody:
        description: The confirmation phrase, plus the optional donate election.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ErasureRequest"
  /api/v1/internal/account/erasure:
    post:
      operationId: requestServiceErasure
      summary: Service-principal erasure trigger, called by the identity deletion worker.
      security:
        - serviceErasureToken: []
      responses:
        "202":
          description: The queued, already-running, or already-completed erasure job.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceErasureAcceptedResponse"
        "401":
          description: Missing, invalid, or expired service token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      description: "NOT a public endpoint and NOT Clerk-authenticated: it authenticates with a signed machine token and is
        excluded from the Clerk middleware so its own guard is the fail-closed enforcement point (ADR-0011). It takes NO
        REQUEST BODY, deliberately — the target owner is the token’s bound claim, so there is nothing for a caller to
        supply and no confirmation phrase (the token IS the authorization). Unlike the user route, an already-erased
        account is an idempotent no-op SUCCESS reporting `status: completed` with no `jobId`, not a `410`: the
        deletion-worker caller only needs to know the erasure is handled. There is no donate election on this path
        (KTD-2) — every owner-only recipe is removed."
  /api/v1/internal/account/food-references:
    post:
      operationId: getServiceFoodReferences
      summary: Which of the erased owner's authored food ids live recipes still reference (U18).
      security:
        - serviceErasureToken: []
      responses:
        "200":
          description: The subset still referenced by live recipes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceFoodReferencesResponse"
        "400":
          description: "The request body, query or path was rejected. `code` says by what: `VALIDATION_FAILED` from the boundary
            parser (with `details.fields` naming each offending field, including a rejected unknown key), or a domain
            code such as `INVALID_VISIBILITY` / `UNKNOWN_INGREDIENT`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No, invalid, expired, or wrong-authorized-party bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: "The caller has an in-flight account erasure, so writes are locked (HAZ-052). `code: ERASURE_IN_PROGRESS`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: "Per-user rate limit exceeded (`code: TOO_MANY_REQUESTS`). ⚠️ This body used to be a bare JSON STRING — the
            throttler exception passed through the filter unchanged — and is now the envelope."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          headers:
            Retry-After:
              description: Seconds until the limit resets.
              schema:
                type: integer
                minimum: 0
                exclusiveMinimum: true
        "500":
          description: Unexpected server error. The body never carries internals.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      description: The deletion worker calls this BETWEEN the food service's erasure begin (tombstone) and its completing
        erasure, deciding Q3b's delete-or-orphan. Machine-token authenticated like the erasure route above; the body
        carries the query's SUBJECT only — the authorization is the verified single-target token, and the published
        schema has no owner field at all.
      requestBody:
        description: The erased owner's authored food ids to check.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceFoodReferencesRequest"
components:
  securitySchemes:
    clerkSessionToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A Clerk session token. Verified networklessly against the instance JWT key, with the signed `azp` checked
        against the service’s authorized-party pattern.
    serviceErasureToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A signed service-principal token, used only by the internal erasure route.
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: {}
      required:
        - code
        - message
      additionalProperties: {}
...
    ServiceFoodReferencesRequest:
      type: object
      properties:
        foodIds:
          maxItems: 10000
          type: array
          items:
            type: string
            minLength: 1
      required:
        - foodIds
      additionalProperties: false
    ServiceFoodReferencesResponse:
      type: object
      properties:
        referencedFoodIds:
          type: array
          items:
            type: string
      required:
        - referencedFoodIds
...
    AccountExport:
      type: object
      properties:
        exportedAt:
          type: string
          minLength: 1
        ownerId:
          type: string
          minLength: 1
        recipes:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                minLength: 1
              ownerId:
                type: string
                minLength: 1
              title:
                type: string
              description:
                nullable: true
                type: string
              prepTimeMinutes:
                type: integer
              cookTimeMinutes:
                type: integer
              totalTimeMinutes:
                type: integer
              servings:
                type: integer
              difficulty:
                nullable: true
                type: string
              mealType:
                nullable: true
                type: string
              averageRating:
                nullable: true
                type: string
              ratingCount:
                type: integer
              visibility:
                type: string
              status:
                type: string
              sourceType:
                type: string
              sourceUrl:
                nullable: true
                type: string
              sourceAttribution:
                nullable: true
                type: string
              clonedFromId:
                nullable: true
                type: string
                minLength: 1
              hasSubstantiveEdit:
                type: boolean
              cuisine:
                nullable: true
                type: string
              dietaryFlags:
                type: array
                items:
                  type: string
              tags:
                type: array
                items:
                  type: string
              authorHandle:
                nullable: true
                type: string
              currentVersion:
                type: integer
              deletedAt:
                nullable: true
                type: string
                minLength: 1
              createdAt:
                type: string
                minLength: 1
              updatedAt:
                type: string
                minLength: 1
            required:
              - id
              - ownerId
              - title
              - description
              - prepTimeMinutes
              - cookTimeMinutes
              - totalTimeMinutes
              - servings
              - difficulty
              - mealType
              - averageRating
              - ratingCount
              - visibility
              - status
              - sourceType
              - sourceUrl
              - sourceAttribution
              - clonedFromId
              - hasSubstantiveEdit
              - cuisine
              - dietaryFlags
              - tags
              - authorHandle
              - currentVersion
              - deletedAt
              - createdAt
              - updatedAt
        collections:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                minLength: 1
              ownerId:
                type: string
                minLength: 1
              name:
                type: string
              description:
                nullable: true
                type: string
              visibility:
                type: string
              sourceCollectionId:
                nullable: true
                type: string
                minLength: 1
              lastPulledAt:
                nullable: true
                type: string
                minLength: 1
              sourceOwnerHandle:
                nullable: true
                type: string
              sourceCollectionName:
                nullable: true
                type: string
              createdAt:
                type: string
                minLength: 1
              updatedAt:
                type: string
                minLength: 1
              recipes:
                type: array
                items:
                  type: object
                  properties:
                    recipeId:
                      type: string
                      minLength: 1
                    addedAt:
                      type: string
                      minLength: 1
                    addedVia:
                      type: string
                  required:
                    - recipeId
                    - addedAt
                    - addedVia
            required:
              - id
              - ownerId
              - name
              - description
              - visibility
              - sourceCollectionId
              - lastPulledAt
              - sourceOwnerHandle
              - sourceCollectionName
              - createdAt
              - updatedAt
              - recipes
        ratings:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                minLength: 1
              recipeId:
                type: string
                minLength: 1
              userId:
                type: string
                minLength: 1
              stars:
                type: integer
              createdAt:
                type: string
                minLength: 1
              updatedAt:
                type: string
                minLength: 1
            required:
              - id
              - recipeId
              - userId
              - stars
              - createdAt
              - updatedAt
        photos:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                minLength: 1
              recipeId:
                type: string
                minLength: 1
              key:
                type: string
                minLength: 1
              url:
                type: string
                format: uri
              thumbnailKey:
                nullable: true
                type: string
                minLength: 1
              thumbnailUrl:
                nullable: true
                type: string
                format: uri
              contentType:
                type: string
              sizeBytes:
                nullable: true
                type: integer
              sortOrder:
                type: integer
              createdAt:
                type: string
                minLength: 1
              updatedAt:
                type: string
                minLength: 1
            required:
              - id
              - recipeId
              - key
              - url
              - thumbnailKey
              - thumbnailUrl
              - contentType
              - sizeBytes
              - sortOrder
              - createdAt
              - updatedAt
        versions:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                minLength: 1
              recipeId:
                type: string
                minLength: 1
              versionNumber:
                type: integer
              baseVersion:
                nullable: true
                type: integer
              s3Key:
                nullable: true
                type: string
                minLength: 1
              createdBy:
                type: string
                minLength: 1
              changeSummary:
                nullable: true
                type: string
              editorHandle:
                nullable: true
                type: string
              createdAt:
                type: string
                minLength: 1
            required:
              - id
              - recipeId
              - versionNumber
              - baseVersion
              - s3Key
              - createdBy
              - changeSummary
              - editorHandle
              - createdAt
        authorHandles:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
                minLength: 1
              displayName:
                type: string
              sourceTimestamp:
                type: string
                minLength: 1
            required:
              - userId
              - displayName
              - sourceTimestamp
      required:
        - exportedAt
        - ownerId
        - recipes
        - collections
        - ratings
        - photos
        - versions
        - authorHandles
    ErasureRequest:
      type: object
      properties:
        confirmationPhrase:
          type: string
          minLength: 1
          maxLength: 100
          description: The exact confirmation phrase. Required. A non-matching value is answered 400 by the service, not by
            request validation — the accepted literal is deliberately not published as an enum.
        publishRecipeIds:
          maxItems: 1000
          type: array
          items:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
      required:
        - confirmationPhrase
      additionalProperties: false
    ErasureRequestAcceptedResponse:
      type: object
      properties:
        jobId:
          type: string
          minLength: 1
        status:
          type: string
          enum:
            - queued
            - running
      required:
        - jobId
        - status
    ServiceErasureAcceptedResponse:
      type: object
      properties:
        jobId:
          type: string
          minLength: 1
        status:
          type: string
          enum:
            - queued
            - running
            - completed
            - failed
        triggerSource:
          type: string
          enum:
            - service
      required:
        - status
        - triggerSource